MAINT: Add provider for paste data
