+++ src/App.tsx
@@ -1,9 +1,14 @@
 import * as React from 'react'
 
+import PasteDataProvider from 'contexts/PasteDataProvider'
 import Root from 'pages/Root'
 
 function App() {
-  return <Root />
+  return (
+    <PasteDataProvider>
+      <Root />
+    </PasteDataProvider>
+  )
 }
 
 export default App
